move cell edit logic to react-bootstrap-table2-editor
