+++ packages/react-bootstrap-table2-editor/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "react-bootstrap-table2-editor",
+  "version": "0.0.1",
+  "description": "it's the editor addon for react-bootstrap-table2",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
