+++ packages/react-bootstrap-table2-editor/src/editor-indicator.js
