+++ packages/react-bootstrap-table2-editor/src/text-editor.js
