+++ packages/react-bootstrap-table2-example/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "bootstrap": "^3.3.7",
     "react-bootstrap-table2": "0.0.1",
+    "react-bootstrap-table2-editor": "0.0.1",
     "react-bootstrap-table2-paginator": "0.0.1",
     "react-bootstrap-table2-overlay": "0.0.1",
     "react-bootstrap-table2-filter": "0.0.1"
