+++ packages/react-bootstrap-table2/src/const.js
@@ -1,15 +1,10 @@
 export default {
   SORT_ASC: 'asc',
   SORT_DESC: 'desc',
-  UNABLE_TO_CELL_EDIT: 'none',
-  CLICK_TO_CELL_EDIT: 'click',
-  DBCLICK_TO_CELL_EDIT: 'dbclick',
-  TIME_TO_CLOSE_MESSAGE: 3000,
   ROW_SELECT_SINGLE: 'radio',
   ROW_SELECT_MULTIPLE: 'checkbox',
   ROW_SELECT_DISABLED: 'ROW_SELECT_DISABLED',
   CHECKBOX_STATUS_CHECKED: 'checked',
   CHECKBOX_STATUS_INDETERMINATE: 'indeterminate',
-  CHECKBOX_STATUS_UNCHECKED: 'unchecked',
-  DELAY_FOR_DBCLICK: 200
+  CHECKBOX_STATUS_UNCHECKED: 'unchecked'
 };
